backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.7"}
psycopg = {extras = ["binary"], version = "^3.2.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.20"
authlib = "^1.4.0"
httpx-oauth = "^0.16.1"
aiortc = "^1.9.0"
aiohttp = "^3.11.11"
websockets = "^12.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
starlette = "^0.45.3"
itsdangerous = "^2.2.0"
pyjwt = "^2.10.1"
stripe = "^7.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
